backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.*
pydantic==2.10.*
pydantic-settings==2.7.*
httpx==0.28.*
groq==0.15.*
sentence-transformers==3.3.*
pymilvus==2.5.*
python-dotenv==1.0.*
bcrypt==4.2.*
